refactor(api): move flatfile store into local backend
Move local transcript/store primitives out of dev-only modules and leave compatibility re-exports for existing fake-headless tests and adapters.
