+++ src/backend/local/LocalBackend.ts
@@ -1,4 +1,5 @@
 import { FakeHeadlessBackend } from "../dev/FakeHeadlessBackend";
+import type { LocalStoreOptions } from "./LocalStore";
 
 export interface LocalBackendOptions {
   storageDir: string;
@@ -7,11 +8,12 @@ export interface LocalBackendOptions {
 
 export class LocalBackend extends FakeHeadlessBackend {
   constructor(options: LocalBackendOptions) {
-    super(options.defaultAgentId, undefined, {
+    const storeOptions: LocalStoreOptions = {
       storageDir: options.storageDir,
       seedDefaultAgent: false,
       strictAgentAccess: true,
       strictConversationAccess: true,
-    });
+    };
+    super(options.defaultAgentId, undefined, storeOptions);
   }
 }
