+++ src/backend/local/index.ts
@@ -4,3 +4,17 @@ export type {
   LocalMessageMetadata,
   LocalMessageProviderMetadata,
 } from "./LocalMessage";
+export {
+  type LocalAgentRecord,
+  LocalBackendNotFoundError,
+  LocalStore,
+  type LocalStoreOptions,
+  type StoredMessage,
+  type StoredTurnInput,
+} from "./LocalStore";
+export type {
+  ProviderStreamPart,
+  ProviderTrajectoryMessage,
+  ProviderTrajectoryRawCapture,
+  ProviderTrajectoryUIMessage,
+} from "./ProviderTrajectory";
